feat(core): change reply uncomputed field policy
Change the policy for uncomputable fields in the reply action:
now these fields are set to the same value than in the request
whereas their were left uninitialized.
This permit to the user to detect them in hook by comparing the
request and the reply.
The fact than the addresses source and destination are only
swapped is not definitive.