feat(udp): add two checkum computation functions
Add a checkcum computation function which process a udp header and
a given ip header (for pseudo header) and the payload.
Add a checksum computation function which takes only the udp header
and find the ip header and the data with the prev/next fields of
netio_header_t.
If it cannot find them, then it returns 0 (which is an impossible
value of checksum according to RFC 768).